[PVR] refactoring: put all pvr and epg shared_ptr typedefs into one header file instead to typedef again and again in several header files.
