+++ xbmc/epg/CMakeLists.txt
@@ -11,6 +11,7 @@ set(HEADERS Epg.h
             EpgDatabase.h
             EpgInfoTag.h
             EpgSearchFilter.h
+            EpgTypes.h
             GUIEPGGridContainer.h
             GUIEPGGridContainerModel.h)
