+++ xbmc/epg/EpgInfoTag.h
@@ -19,15 +19,18 @@
  *
  */
 
-#include <memory>
-#include <string>
-#include <vector>
-
 #include "XBDateTime.h"
 #include "addons/kodi-addon-dev-kit/include/kodi/xbmc_pvr_types.h"
+#include "pvr/PVRTypes.h"
 #include "pvr/channels/PVRChannel.h"
+#include "pvr/recordings/PVRRecording.h"
 #include "pvr/timers/PVRTimerInfoTag.h"
 #include "utils/ISerializable.h"
+
+#include "epg/EpgTypes.h"
+
+#include <string>
+#include <vector>
 
 #define EPG_DEBUGGING 0
 
@@ -36,9 +39,6 @@ class CVariant;
 namespace EPG
 {
   class CEpg;
-
-  class CEpgInfoTag;
-  typedef std::shared_ptr<EPG::CEpgInfoTag> CEpgInfoTagPtr;
 
   class CEpgInfoTag : public ISerializable
   {
